+++ src/IdentityAccess/Contract/UserContact.php
@@ -1,0 +1,13 @@
+<?php
+
+/*
+ * SPDX-FileCopyrightText: 2026 The Kadupul project and contributors
+ * SPDX-License-Identifier: GPL-3.0-or-later
+ */
+
+namespace Kadupul\IdentityAccess\Contract;
+
+final readonly class UserContact
+{
+    public function __construct(public string $email, public string $name) {}
+}